fix(v2.10/specs): close audit's spec-shape gaps — multisig member list + lending liquidation gate
Audit's two remaining HIGHs were spec-shape gaps the codegen couldn't
synthesize without more model: multisig modeled membership as a count
rather than a list, and lending's liquidate had no health threshold.
Both fixed with surgical spec edits + supporting codegen polish.

Spec edits:

- multisig.qedspec — `State.Active` now carries `members : Map[32]
  Pubkey` and `voted : Map[32] U8`. `approve` and `reject` add two
  new `requires` clauses: `state.members[member_index] == approver`
  (binds index to signer) and `state.voted[member_index] == 0`
  (per-member dedup). Effects flip `voted[member_index] := 1`.
- lending.qedspec — `liquidate` now requires `state.amount >
  state.collateral else AccountHealthy`. `Error` block adds
  `AccountHealthy`.

Supporting codegen:

- `find_state_account` now disambiguates multi-state handlers via
  `handler.on_account` — lending::liquidate (writable PDAs `loan` AND
  `pool`) picks `loan` since the lifecycle is `Loan.Active ->
  Loan.Liquidated`. Without this, `s.amount > s.collateral` survived
  un-rewritten in guards.rs.
- `bind_state` now also rewrites bare handler-account idents (e.g.
  `approver` in `state.members[i] == approver`) to runtime address
  loads `(*ctx.<name>.to_account_view().address())`. Previously these
  bare idents resolved to nothing in the guards.rs scope.
- `path_to_rust` Index emission casts to `usize`: `voted[member_index]`
  now lowers to `voted[(member_index) as usize]` so a u8/u16/Fin
  handler param indexes a `[T; N]` array. Effect-LHS render keeps the
  raw index because the Lean side reads the same field string.

Test routing change:

- The bundled `multisig.qedspec` now uses Map fields, which routes
  lean_gen through `render_indexed_state` — that path deliberately
  doesn't emit theorem bodies (proofs live in user-owned
  `Proofs.lean`). Six `lean_gen` tests pinned to the legacy auto-prove
  path now read a frozen scalar-only `MULTISIG_SCALAR_FIXTURE` so the
  assertions stay meaningful while the real fixture evolves.

All three examples still compile cleanly. All 435 qedgen tests pass.
Audit's two HIGH findings closed.

+++ examples/rust/percolator/programs/src/guards.rs
@@ -17,9 +17,9 @@ pub fn add_user<'info>(ctx: &mut AddUser<'info>, i: usize) -> Result<(), Program
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 0
-    if !(ctx.vault.accounts[i].active == 0) { return Err(ProgramError::from(PercolatorError::SlotAlreadyActive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 0) { return Err(ProgramError::from(PercolatorError::SlotAlreadyActive)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int))
-    if !(((ctx.vault.accounts[i].capital.get()) as i128) + ctx.vault.accounts[i].pnl.get() >= ((0) as i128)) { return Err(ProgramError::from(PercolatorError::AccountUnderwater)); }
+    if !(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128)) { return Err(ProgramError::from(PercolatorError::AccountUnderwater)); }
     Ok(())
 }
 
@@ -29,9 +29,9 @@ pub fn add_lp<'info>(ctx: &mut AddLp<'info>, i: usize) -> Result<(), ProgramErro
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 0
-    if !(ctx.vault.accounts[i].active == 0) { return Err(ProgramError::from(PercolatorError::SlotAlreadyActive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 0) { return Err(ProgramError::from(PercolatorError::SlotAlreadyActive)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int))
-    if !(((ctx.vault.accounts[i].capital.get()) as i128) + ctx.vault.accounts[i].pnl.get() >= ((0) as i128)) { return Err(ProgramError::from(PercolatorError::AccountUnderwater)); }
+    if !(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128)) { return Err(ProgramError::from(PercolatorError::AccountUnderwater)); }
     Ok(())
 }
 
@@ -41,13 +41,13 @@ pub fn reclaim_empty_account<'info>(ctx: &mut ReclaimEmptyAccount<'info>, i: usi
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: s.accounts[i].capital = 0
-    debug_assert!(ctx.vault.accounts[i].capital.get() == 0);
+    debug_assert!(ctx.vault.accounts[(i) as usize].capital.get() == 0);
     // requires: s.accounts[i].reserved_pnl = 0
-    debug_assert!(ctx.vault.accounts[i].reserved_pnl.get() == 0);
+    debug_assert!(ctx.vault.accounts[(i) as usize].reserved_pnl.get() == 0);
     // requires: s.accounts[i].fee_credits = 0
-    debug_assert!(ctx.vault.accounts[i].fee_credits.get() == 0);
+    debug_assert!(ctx.vault.accounts[(i) as usize].fee_credits.get() == 0);
     Ok(())
 }
 
@@ -57,9 +57,9 @@ pub fn close_account<'info>(ctx: &mut CloseAccount<'info>, i: usize) -> Result<(
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: s.V ≥ s.accounts[i].capital
-    if !(ctx.vault.V.get() >= ctx.vault.accounts[i].capital.get()) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
+    if !(ctx.vault.V.get() >= ctx.vault.accounts[(i) as usize].capital.get()) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
     Ok(())
 }
 
@@ -69,7 +69,7 @@ pub fn deposit<'info>(ctx: &mut Deposit<'info>, i: usize, amount: u128) -> Resul
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: s.V + amount ≤ 10000000000000000
     if !(ctx.vault.V.get() + amount <= 10000000000000000) { return Err(ProgramError::from(PercolatorError::VaultOverflow)); }
     Ok(())
@@ -81,11 +81,11 @@ pub fn withdraw<'info>(ctx: &mut Withdraw<'info>, i: usize, amount: u128) -> Res
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: s.accounts[i].capital ≥ amount
-    if !(ctx.vault.accounts[i].capital.get() >= amount) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
+    if !(ctx.vault.accounts[(i) as usize].capital.get() >= amount) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((amount) : Int))
-    if !(((ctx.vault.accounts[i].capital.get()) as i128) + ctx.vault.accounts[i].pnl.get() >= ((amount) as i128)) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
+    if !(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((amount) as i128)) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
     Ok(())
 }
 
@@ -105,7 +105,7 @@ pub fn deposit_fee_credits<'info>(ctx: &mut DepositFeeCredits<'info>, i: usize, 
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: s.V + amount ≤ 10000000000000000
     if !(ctx.vault.V.get() + amount <= 10000000000000000) { return Err(ProgramError::from(PercolatorError::VaultOverflow)); }
     Ok(())
@@ -117,9 +117,9 @@ pub fn convert_released_pnl<'info>(ctx: &mut ConvertReleasedPnl<'info>, i: usize
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: s.accounts[i].reserved_pnl ≥ x
-    if !(ctx.vault.accounts[i].reserved_pnl.get() >= x) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
+    if !(ctx.vault.accounts[(i) as usize].reserved_pnl.get() >= x) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
     // requires: s.V ≥ x
     if !(ctx.vault.V.get() >= x) { return Err(ProgramError::from(PercolatorError::InsufficientFunds)); }
     Ok(())
@@ -131,9 +131,9 @@ pub fn execute_trade<'info>(ctx: &mut ExecuteTrade<'info>, a: usize, b: usize, s
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[a].active = 1
-    if !(ctx.vault.accounts[a].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(a) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: s.accounts[b].active = 1
-    if !(ctx.vault.accounts[b].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(b) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: a ≠ b
     debug_assert!(a != b);
     // requires: (((size_q) * ((((exec_price) : Int)))) / (1000000)) ≤ (((100000000000000000000) : Int))
@@ -147,9 +147,9 @@ pub fn liquidate_case_0<'info>(ctx: &mut LiquidateCase0<'info>, i: usize) -> Res
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int))
-    debug_assert!(((ctx.vault.accounts[i].capital.get()) as i128) + ctx.vault.accounts[i].pnl.get() >= ((0) as i128));
+    debug_assert!(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128));
     // requires: 0 = 1
     if !(false) { return Err(ProgramError::from(PercolatorError::AccountHealthy)); }
     Ok(())
@@ -161,11 +161,11 @@ pub fn liquidate_case_1<'info>(ctx: &mut LiquidateCase1<'info>, i: usize) -> Res
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: ¬((((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int)))
-    debug_assert!(!(((ctx.vault.accounts[i].capital.get()) as i128) + ctx.vault.accounts[i].pnl.get() >= ((0) as i128)));
+    debug_assert!(!(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128)));
     // requires: (((s.accounts[i].capital) : Int)) + s.accounts[i].pnl + (((s.I) : Int)) ≥ (((0) : Int))
-    debug_assert!(((ctx.vault.accounts[i].capital.get()) as i128) + ctx.vault.accounts[i].pnl.get() + ((ctx.vault.I.get()) as i128) >= ((0) as i128));
+    debug_assert!(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() + ((ctx.vault.I.get()) as i128) >= ((0) as i128));
     Ok(())
 }
 
@@ -175,11 +175,11 @@ pub fn liquidate_otherwise<'info>(ctx: &mut LiquidateOtherwise<'info>, i: usize)
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     // requires: ¬((((s.accounts[i].capital) : Int)) + s.accounts[i].pnl ≥ (((0) : Int)))
-    debug_assert!(!(((ctx.vault.accounts[i].capital.get()) as i128) + ctx.vault.accounts[i].pnl.get() >= ((0) as i128)));
+    debug_assert!(!(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() >= ((0) as i128)));
     // requires: ¬((((s.accounts[i].capital) : Int)) + s.accounts[i].pnl + (((s.I) : Int)) ≥ (((0) : Int)))
-    debug_assert!(!(((ctx.vault.accounts[i].capital.get()) as i128) + ctx.vault.accounts[i].pnl.get() + ((ctx.vault.I.get()) as i128) >= ((0) as i128)));
+    debug_assert!(!(((ctx.vault.accounts[(i) as usize].capital.get()) as i128) + ctx.vault.accounts[(i) as usize].pnl.get() + ((ctx.vault.I.get()) as i128) >= ((0) as i128)));
     // requires: 0 = 1
     if !(false) { return Err(ProgramError::from(PercolatorError::BankruptPosition)); }
     Ok(())
@@ -191,7 +191,7 @@ pub fn settle_account<'info>(ctx: &mut SettleAccount<'info>, i: usize) -> Result
     // lifecycle: require status == Active
     if ctx.vault.status != Status::Active as u8 { return Err(ProgramError::from(PercolatorError::InvalidLifecycle)); }
     // requires: s.accounts[i].active = 1
-    if !(ctx.vault.accounts[i].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
+    if !(ctx.vault.accounts[(i) as usize].active == 1) { return Err(ProgramError::from(PercolatorError::SlotInactive)); }
     Ok(())
 }
